Move createTestPoller tests from internal to public (core-lro)
Move 5 tests that only use createTestPoller (public API) from
test/internal/poller-operation.spec.ts to test/public/poller.spec.ts.
Internal file retains 6 tests that use internal APIs (deserializeState,
pollOperation, initOperation, buildCreatePoller).

All 347 tests pass with build-test clean.
